fix(core): moved a specific use-case func out of s3
refs. metriport/metriport-internal#1905
