+++ packages/core/src/external/aws/s3.ts
@@ -24,7 +24,7 @@ const defaultS3RetriesConfig = {
   initialDelay: 500,
 };
 
-async function executeWithRetriesS3<T>(
+export async function executeWithRetriesS3<T>(
   fn: () => Promise<T>,
   options?: ExecuteWithRetriesOptions<T>
 ): Promise<T> {
@@ -354,41 +354,6 @@ export class S3Utils {
       console.error(`Error during file deletion: ${JSON.stringify(error)}`);
       throw error;
     }
-  }
-
-  async retrieveDocumentIdsFromS3(
-    cxId: string,
-    patientId: string,
-    bucketName: string
-  ): Promise<string[]> {
-    const Prefix = `${cxId}/${patientId}/uploads/`;
-
-    const params = {
-      Bucket: bucketName,
-      Prefix,
-    };
-
-    const data = await executeWithRetriesS3(() => this._s3.listObjectsV2(params).promise());
-    const documentContents = (
-      await Promise.all(
-        data.Contents?.filter(item => item.Key && item.Key.endsWith("_metadata.xml")).map(
-          async item => {
-            if (item.Key) {
-              const params = {
-                Bucket: bucketName,
-                Key: item.Key,
-              };
-
-              const data = await executeWithRetriesS3(() => this._s3.getObject(params).promise());
-              return data.Body?.toString();
-            }
-            return undefined;
-          }
-        ) || []
-      )
-    ).filter((item): item is string => Boolean(item));
-
-    return documentContents;
   }
 
   async listObjects(bucket: string, prefix: string): Promise<AWS.S3.ObjectList | undefined> {
