+++ packages/core/src/external/carequality/dq/process-inbound-dq.ts
@@ -1,11 +1,11 @@
 import { InboundDocumentQueryReq, InboundDocumentQueryResp } from "@metriport/ihe-gateway-sdk";
 import { executeWithNetworkRetries } from "@metriport/shared";
 import axios from "axios";
-import { CCD_SUFFIX, createUploadFilePath } from "../../../domain/document/upload";
+import { CCD_SUFFIX, UPLOADS_FOLDER, createUploadFilePath } from "../../../domain/document/upload";
 import { Config } from "../../../util/config";
 import { out } from "../../../util/log";
 import { capture } from "../../../util/notifications";
-import { S3Utils } from "../../aws/s3";
+import { S3Utils, executeWithRetriesS3 } from "../../aws/s3";
 import {
   IHEGatewayError,
   XDSRegistryError,
@@ -14,6 +14,7 @@ import {
 } from "../error";
 import { validateBasePayload } from "../shared";
 import { decodePatientId } from "./utils";
+import { createFolderName } from "../../../domain/filename";
 
 const region = Config.getAWSRegion();
 const s3Utils = new S3Utils(region);
@@ -77,12 +78,48 @@ export async function processInboundDocumentQuery(
 }
 
 async function getDocumentContents(cxId: string, patientId: string): Promise<string[]> {
-  const documentContents = await s3Utils.retrieveDocumentIdsFromS3(cxId, patientId, bucket);
+  const documentContents = await retrieveXmlContentsFromMetadataFilesOnS3(cxId, patientId, bucket);
 
   if (!documentContents.length) {
     const msg = `Error getting document contents`;
     capture.error(msg, { extra: { cxId, patientId } });
     throw new XDSRegistryError("Internal Server Error");
   }
+  return documentContents;
+}
+
+async function retrieveXmlContentsFromMetadataFilesOnS3(
+  cxId: string,
+  patientId: string,
+  bucketName: string
+): Promise<string[]> {
+  const folderName = createFolderName(cxId, patientId);
+  const Prefix = `${folderName}/${UPLOADS_FOLDER}/`;
+
+  const params = {
+    Bucket: bucketName,
+    Prefix,
+  };
+
+  const data = await executeWithRetriesS3(() => s3Utils._s3.listObjectsV2(params).promise());
+  const documentContents = (
+    await Promise.all(
+      data.Contents?.filter(item => item.Key && item.Key.endsWith("_metadata.xml")).map(
+        async item => {
+          if (item.Key) {
+            const params = {
+              Bucket: bucketName,
+              Key: item.Key,
+            };
+
+            const data = await executeWithRetriesS3(() => s3Utils._s3.getObject(params).promise());
+            return data.Body?.toString();
+          }
+          return undefined;
+        }
+      ) || []
+    )
+  ).filter((item): item is string => Boolean(item));
+
   return documentContents;
 }
